Added lots of styling and layout changes
