+++ uofg_prepare/package.json
@@ -12,7 +12,8 @@
     "@mantine/core": "^7.3.2",
     "next": "14.0.4",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "sass": "^1.69.7"
   },
   "devDependencies": {
     "@types/node": "^20",
