+++ uofg_prepare/src/components/shared/PageHeadingText.tsx
@@ -9,7 +9,7 @@ export const PageHeadingText = ({ children }: {children: ReactNode}) => {
 	return (
 		<>
 			<Head>
-				<title>{`Backstage - ${children}`}</title>
+				<title>{`UofG Prepare - ${children}`}</title>
 			</Head>
 			<Title order={1} size="h3" w={500}>
 				{children}
