+++ uofg_prepare/src/styles/Layout.module.css
@@ -1,0 +1,3 @@
+.header {
+    background-color: var(--mantine-color-blue-9);
+}
