+++ uofg_prepare/src/styles/NavbarButton.module.css
@@ -7,23 +7,23 @@
     margin-top: 4px;
     margin-bottom: 4px;
     padding: var(--spacing-xs) var(--spacing-md);
-    color: var(--button-text-color);
+    color: var(--mantine-color-blue-9);
     font-size: var(--font-size-sm);
     border-radius: var(--radius-sm);
   
     .navbar-icon {
-      color: var(--navbar-icon-color);
+      color: var(--mantine-color-blue);
       display: flex;
       align-items: center;
     }
   }
   
   .active {
-    background-color: var(--active-bg);
-    color: var(--active-text-color);
+    background-color: var(--mantine-color-gray-3);
+    color: var(--mantine-color-blue-9);
   
     .navbar-icon {
-      color: var(--navbar-icon-active-color);
+      color: var(--mantine-color-white);
     }
   }
   
@@ -39,8 +39,8 @@
     width: auto;
     border-radius: 0 var(--radius-sm) var(--radius-sm) 0;
     font-size: var(--font-size-sm);
-    color: var(--link-color);
-    border-left: 1px solid var(--link-border-color);
+    color: var(--mantine-color-blue-9);
+    border-left: 1px solid var(--mantine-color-blue-9);
   }
   
   .chevron {
