+++ uofg_prepare/src/styles/Tabs.module.css
@@ -1,0 +1,38 @@
+.tab {
+  position: relative;
+  border: 1px solid light-dark(var(--mantine-color-gray-2), var(--mantine-color-dark-4));
+  background-color: light-dark(var(--mantine-color-white), var(--mantine-color-dark-6));
+  transition: background-color 0.3s ease;
+}
+
+.tab:first-of-type {
+  border-radius: 4px 0 0 4px;
+}
+
+.tab:last-of-type {
+  border-radius: 0 4px 4px 0;
+}
+
+.tab + .tab {
+  border-left-width: 0;
+}
+
+.tab[data-rtl='true'] {
+  border-right-width: 0;
+  border-left-width: 1px;
+}
+
+.tab:hover {
+  background-color: light-dark(var(--mantine-color-gray-0), var(--mantine-color-dark-5));
+}
+
+.tab[data-active] {
+  z-index: 1;
+  background-color: var(--mantine-color-blue-9);
+  border-color: var(--mantine-color-blue-9);
+  color: var(--mantine-color-white);
+}
+
+.tab[data-active]:hover {
+  background-color: var(--mantine-color-blue-9);
+}
